requirements.txt
django==5.0.1
djangorestframework=3.0.1
asgiref==3.7.2
pytz==2023.3.post1
sqlparse==0.4.4